AhtoHotKey-time-stamp.ahk
This version handles time zones a little better. When using cc or ccc if your computer is not set to central time zone, it will append a single letter to the end of the time to denote the timezone. This will only handle times within the contiguous US and Alaska.